zfill implemented add arg to script to keep binary

+++ tests/run.py
@@ -2,6 +2,18 @@ from colorama import Fore, Style
 from glob import glob
 import subprocess
 import os
+import sys
+
+tests = ["memory",
+	"conversions",
+	"chartype",
+	"strcat",
+	"strset",
+	"swapcase",
+	"capitalize",
+	"zfill",
+	"chained_lst",
+	"all"]
 
 def run_tests(name: str):
 	binary = name + "_tests"
@@ -42,10 +54,9 @@ def run_tests(name: str):
 			print(Fore.LIGHTRED_EX + "Leaks were found!    " + Style.RESET_ALL)
 	except:
 		print(Fore.LIGHTRED_EX + "Failed to run Valgrind!" + Style.RESET_ALL)
-	os.remove(binary)
+	if len(sys.argv) == 1:
+		os.remove(binary)
 print(Fore.LIGHTGREEN_EX + "Welcome to neo_libft tests !" + Style.RESET_ALL)
-
-tests = ["memory", "conversions", "chartype", "strcat", "strset", "swapcase", "capitalize", "chained_lst", "all"]
 tests.append("exit")
 for index, test in enumerate(tests):
 	print(Fore.LIGHTMAGENTA_EX + "{}: ".format(index) + Fore.LIGHTYELLOW_EX + test + Style.RESET_ALL)
